Added a proof-of-concept for buttons, changed variables to adhere to Python naming conventions, and formatted comments a bit more nicely
This edit of the code serves as a prototype for the clicker game, and a proof-of-concept that buttons are working as intended. Many customization options are available (by changing some of the constants/ variables in all-caps).
All prototype code is identified by the phrase "PROTOTYPE CODE" in a comment above the respective code blocks.
Many variables were changed to adhere to the Python naming conventions, mainly the convention that constants must be in all-caps.
Comments were added, some in a header fashion, others to help the readability of the code.