Added a general system for button creation and registering clicks to main.py
All the proof-of-concept prototype code has been removed. There is now a barebones demonstration of how the clicker will be implemented.
There are now functions that create buttons automatically and append them to a button dictionary, which contains the render of the button and its coordinates. There is now also a working counter to keep track of clicks.