Game is now titled "Medieval Munchies". Updated main.py to now use a main() function to run the game loop, updated the main menu UI, and added icon graphics to the game
The game has been renamed to Medieval Munchies for the time being.
The create_button() method has been renamed to create_text, as we have removed button functionality for the time being. Function description was updated accordingly.
There now exists a create_image() method that supports importing images from our Images folder. This function has the ability to scale an image (in pixels) up or down, and may have rotation functionality in the near future.
The entirety of the main user interface has now been bundled into a function called main_menu(), which simply instantiates all of the text required for the main menu.
Currently, sprites exist for the Carrot Sword, Anvil, Shopping Cart, and Chef Character in this build of the project.